==> dataset/code_for_managing_dataset/1_dataset_distribution_code.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

df1 = pd.read_csv('/kaggle/input/news-article-data-set-from-indian-express/News_Articles_Indian_Express.csv')
df2 = pd.read_csv('//kaggle/input/cnn-news18-headlines-with-descriptions/cnn_news18_headlines.csv')
df3 = pd.read_csv('/kaggle/input/news-summary/news_summaries.csv')

<h2>Dataset 1</h2>

df1.head()

df1.describe(include='object')

df1.drop(columns=['Unnamed: 0', 'article_id', 'date', 'article_type', 'article_length', 'desc', 'url'], inplace=True)

df1.rename(columns={'headline':'title', 'articles':'content'}, inplace=True)

df1.head()

df1.describe(include='object')

<h2>Second dataset</h2>

df2.head()

df2.describe(include='object')

df2.drop(columns=['Unnamed: 0', 'date'], inplace=True)

df2.rename(columns={'headline':'title', 'description':'content'}, inplace=True)

df2.head()

df2.describe(include='object')

<h2>Dataset 3</h2>

df3.head()

df3.describe(include='object')

df3.drop(columns=['date', 'author', 'read_more'], inplace=True)

df3.rename(columns={'headlines':'title', 'text':'content'}, inplace=True)

df3.head()

df3.describe(include='object')

Now merging all the three datatframes

df = pd.concat([df1, df2, df3], ignore_index=True)

for i in range(0, len(df)):
    df.iloc[i, 1] = df.iloc[i, 1][:550]

df.head()

df.describe(include='object')

df = df.drop_duplicates(subset='title')

df.insert(0, 'labels', '')

df.describe(include='object')

df.head()

<b>So, `36051/3 = 12017` for each person</b>

df_aanshu = df[:12017]
df_vansh = df[12017:24034]
df_twisha = df[24034:]

df_aanshu.describe(include='object')

df_vansh.describe(include='object')

df_twisha.describe(include='object')

df_aanshu.to_csv('aanshu.csv')
df_vansh.to_csv('vansh.csv')
df_twisha.to_csv('twisha.csv')

<h2>Making labels of dataset</h2>

ministry_list = ['Ministry of Railways',
                 'Ministry of Rural Development',
                 'Ministry of Steel',
                 'Ministry of Science & Technology',
                 'Ministry of Information & Broadcasting',
                 'Ministry of Food Processing Industries',
                 'Ministry of Health and Family Welfare',
                 'Ministry of Human Resource Development',
                 'Ministry of Agriculture',
                 'Ministry of Environment and Forests',
                 'Ministry of Chemicals and Fertilizers',
                 'Ministry of Water Resources',
                 'Ministry of Defence',
                 'Ministry of Petroleum & Natural Gas',
                 "President's Secretariat",
                 'Ministry of Micro,Small & Medium Enterprises',
                 'Ministry of Mines',
                 'Ministry of Tourism',
                 'Ministry of Housing & Urban Affairs',
                 'Ministry of Coal',
                 "Prime Minister's Office",
                 'Ministry of Textiles',
                 'Ministry of Commerce & Industry',
                 'Ministry of External Affairs',
                 'Ministry of Social Justice & Empowerment',
                 'Ministry of Power',
                 'Ministry of Consumer Affairs, Food & Public Distribution',
                 'Ministry of Heavy Industries & Public Enterprises',
                 'Ministry of Communications',
                 'Ministry of Shipping',
                 'Ministry of Finance',
                 'Ministry of Tribal Affairs',
                 'Ministry of Statistics & Programme Implementation',
                 'Ministry of Labour & Employment',
                 'Ministry of Law & Justice',
                 "Vice President's Secretariat",
                 'Ministry of Civil Aviation',
                 'Ministry for Development of North-East Region',
                 'UPSC',
                 'Ministry of Agro & Rural Industries',
                 'Ministry of Home Affairs',
                 'Ministry of Youth Affairs and Sports',
                 'Special Service and Features',
                 'Ministry of New and Renewable Energy',
                 'Ministry of Parliamentary Affairs',
                 'Planning Commission',
                 'Ministry of Personnel, Public Grievances & Pensions',
                 'Election Commission',
                 'Department of Space',
                 'Ministry of Disinvestment',
                 'Department of Ocean Development',
                 'Ministry of Overseas Indian Affairs',
                 'Ministry of Housing and Urban Poverty Alleviation',
                 'Ministry of Culture',
                 'Ministry of Company Affairs',
                 'Ministry of Panchayati Raj',
                 'Cabinet Committee on Economic Affairs (CCEA)',
                 'Cabinet',
                 'Department of Atomic Energy',
                 'Cabinet Committee Decisions',
                 'No Ministry'
                ]

df_ministry_label = pd.DataFrame(ministry_list)

df_ministry_label.head()

# df_ministry_label['label'] =range(1, len(df_ministry_label) + 1)

df_ministry_label.head()

df_ministry_label.tail()

df_ministry_label.to_csv('ministry_label.csv')
