==> dataset/code_for_managing_dataset/2_preProcessing_of_df_and_combining.ipynb <==
import pandas as pd

# Define file paths
file1_path = "Aansu.csv"
file2_path = "Vansh_final_labelled.csv"
file3_path = "twisha.csv"

df1 = pd.read_csv(file1_path)

df2 = pd.read_csv(file2_path)

df3 = pd.read_csv(file3_path)

df1.head(3)

df1.describe()

df1.describe(include='object')

df2.head(3)

df2.describe()

df3.head(3)

df3.describe()

def replace_nan_to_space(df, columns):
  """
  Replaces NaN values in specified columns with spaces, while leaving other values untouched.

  Args:
    df: The Pandas DataFrame.
    columns: A list of column names to apply the replacement to.

  Returns:
    The modified Pandas DataFrame.
  """
  for col in columns:
    mask = df[col].notnull()
    df.loc[mask, col] = df.loc[mask, col]  # Avoid unnecessary changes
    df.loc[~mask, col] = ' '
  return df

updated_df1 = replace_nan_to_space(df1, ["Unnamed: 4", "Unnamed: 5", "Unnamed: 6"])

updated_df1.head(3)

# Create new column with combined values
updated_df1["content"] = updated_df1["content"] + updated_df1["Unnamed: 4"].map(str) + updated_df1["Unnamed: 5"].map(str) + updated_df1["Unnamed: 6"].map(str)
# Drop multiple columns
updated_df1 = updated_df1.drop(['Unnamed: 4', 'Unnamed: 5', 'Unnamed: 6'], axis=1)

updated_df1.head(3)

updated_df2 = replace_nan_to_space(df2, ["Unnamed: 4", "Unnamed: 5", "Unnamed: 6"])

# Create new column with combined values
updated_df2["content"] = updated_df2["content"] + updated_df2["Unnamed: 4"].map(str) + updated_df2["Unnamed: 5"].map(str) + updated_df2["Unnamed: 6"].map(str)
# Drop multiple columns
updated_df2 = updated_df2.drop(['Unnamed: 4', 'Unnamed: 5', 'Unnamed: 6'], axis=1)

updated_df2.head(3)

updated_df3 = replace_nan_to_space(df3, ["Unnamed: 4"])

updated_df3.head(3)

# Create new column with combined values
updated_df3["content"] = updated_df3["content"] + updated_df3["Unnamed: 4"].map(str)
# Drop multiple columns
updated_df3 = updated_df3.drop(['Unnamed: 4'], axis=1)

updated_df3.head(3)

# Output file path
output_path = "combined_dataset.csv"

# Combine DataFrames
combined_df = pd.concat([updated_df1, updated_df2, updated_df3], ignore_index=True)

# Save combined DataFrame to CSV
combined_df.to_csv(output_path, index=False)

print(f"Successfully combined CSV files into '{output_path}'")
